moved endpoints to moth controller
